Makefile: Disable "format-truncation" and "unused-constant-variable" warning globally
This was introduced in GCC 7.x: https://gcc.gnu.org/onlinedocs/gcc/Warning-Options.html#index-Wformat-truncation

This warning doesn't really make sense in the context of the kernel
because there is no way for the values it complains about to be cut
off. For example, it complains about a cpu int getting truncated,
when in reality it will never be over 7 for this device (octacore).

Also reorganize a bit.

https://github.com/JDCTeam/platform_kernel_samsung_jf/commit/2306df7dba74040a9be69fb5d1a81701907cd670
https://github.com/JDCTeam/platform_kernel_samsung_jf/commit/96c3f1752ca41676cb618acfeaeaa060a7c0abc9