ext4: fix check to prevent initializing reserved inodes
Patch 240030: "ext4: only look at the bg_flags field if it is
valid" will complain if block group zero does not have the
EXT4_BG_INODE_ZEROED flag set. Unfortunately, this is not correct,
since a freshly created file system has this flag cleared. It gets
almost immediately after the file system is mounted read-write --- but
the following somewhat unlikely sequence will end up triggering a
false positive report of a corrupted file system:

	mkfs.ext4 /dev/vdc
	mount -o ro /dev/vdc /vdc
	mount -o remount,rw /dev/vdc

Instead, when initializing the inode table for block group zero, test
to make sure that itable_unused count is not too large, since that is
the case that will result in some or all of the reserved inodes
getting cleared.

This fixes the failures reported by Eric Whiteney when running
generic/230 and generic/231 in the the nojournal test case.

Change-Id: Ifc5c9b6ae15d91b098bd76d16b22827c3e30da8f
Fixes: patch 240030 ("ext4: only look at the bg_flags field if it is valid")